Add files via upload
This python script was completely written by ChatGPT Dec 15 Version. The user specifies a folder and it searches for all ".mp4", ".avi", ".mkv", and ".mov" files within that folder and creates a simple HTML file with hyperlinks to those file paths. this can be useful for locating certain files that are lost deep within the vast span of folder trees on a hard drive. I used this to find old videos that I had tucked away somewhere and forgotten, but you can modify it to find any file type you like or expand upon it for more detailed functionality and style. You will need the 'Humanize' python package installed as well to run this script.